added the feature handler at backend and made the changes in the frontend
