+++ core/controllers/admin.py
@@ -39,7 +39,6 @@ from core.domain import exp_fetchers
 from core.domain import exp_services
 from core.domain import fs_services
 from core.domain import opportunity_services
-from core.domain import platform_feature_services as feature_services
 from core.domain import platform_parameter_domain as parameter_domain
 from core.domain import question_domain
 from core.domain import question_services
@@ -111,7 +110,6 @@ class AdminHandlerNormalizePayloadDict(TypedDict):
     config_property_id: Optional[str]
     data: Optional[str]
     topic_id: Optional[str]
-    feature_name: Optional[str]
     commit_message: Optional[str]
     new_rules: Optional[List[parameter_domain.PlatformParameterRule]]
     exp_id: Optional[str]
@@ -139,7 +137,6 @@ class AdminHandler(
                         'save_config_properties', 'revert_config_property',
                         'upload_topic_similarities',
                         'regenerate_topic_related_opportunities',
-                        'update_feature_flag_rules',
                         'rollback_exploration_to_safe_state'
                     ]
                 },
@@ -198,12 +195,6 @@ class AdminHandler(
                 },
                 'default_value': None
             },
-            'feature_name': {
-                'schema': {
-                    'type': 'basestring'
-                },
-                'default_value': None
-            },
             'commit_message': {
                 'schema': {
                     'type': 'basestring'
@@ -238,8 +229,6 @@ class AdminHandler(
         topic_summary_dicts = [
             summary.to_dict() for summary in topic_summaries]
 
-        feature_flag_dicts = feature_services.get_all_feature_flag_dicts()
-
         config_properties = config_domain.Registry.get_config_property_schemas()
         # Removes promo-bar related configs as promo-bar is handled by
         # release coordinators in /release-coordinator page.
@@ -260,7 +249,6 @@ class AdminHandler(
             'human_readable_roles': role_services.HUMAN_READABLE_ROLES,
             'role_to_actions': role_services.get_role_actions(),
             'topic_summaries': topic_summary_dicts,
-            'feature_flags': feature_flag_dicts,
         })
 
     @acl_decorators.can_access_admin_page
@@ -367,7 +355,13 @@ class AdminHandler(
                 result = {
                     'opportunities_count': opportunities_count
                 }
-            elif action == 'rollback_exploration_to_safe_state':
+            else:
+                # The handler schema defines the possible values of 'action'.
+                # If 'action' has a value other than those defined in the
+                # schema, a Bad Request error will be thrown. Hence, 'action'
+                # must be 'rollback_exploration_to_safe_state' if this branch is
+                # executed.
+                assert action == 'rollback_exploration_to_safe_state'
                 exp_id = self.normalized_payload.get('exp_id')
                 if exp_id is None:
                     raise Exception(
@@ -379,45 +373,6 @@ class AdminHandler(
                 result = {
                     'version': version
                 }
-            else:
-                # The handler schema defines the possible values of 'action'.
-                # If 'action' has a value other than those defined in the
-                # schema, a Bad Request error will be thrown. Hence, 'action'
-                # must be 'update_feature_flag_rules' if this branch is
-                # executed.
-                assert action == 'update_feature_flag_rules'
-                feature_name = self.normalized_payload.get('feature_name')
-                if feature_name is None:
-                    raise Exception(
-                        'The \'feature_name\' must be provided when the action'
-                        ' is update_feature_flag_rules.'
-                    )
-                new_rules = self.normalized_payload.get('new_rules')
-                if new_rules is None:
-                    raise Exception(
-                        'The \'new_rules\' must be provided when the action'
-                        ' is update_feature_flag_rules.'
-                    )
-                commit_message = self.normalized_payload.get('commit_message')
-                if commit_message is None:
-                    raise Exception(
-                        'The \'commit_message\' must be provided when the '
-                        'action is update_feature_flag_rules.'
-                    )
-
-                try:
-                    feature_services.update_feature_flag_rules(
-                        feature_name, self.user_id, commit_message,
-                        new_rules)
-                except (
-                        utils.ValidationError,
-                        feature_services.FeatureFlagNotFoundException) as e:
-                    raise self.InvalidInputException(e)
-
-                new_rule_dicts = [rules.to_dict() for rules in new_rules]
-                logging.info(
-                    '[ADMIN] %s updated feature %s with new rules: '
-                    '%s.' % (self.user_id, feature_name, new_rule_dicts))
             self.render_json(result)
         except Exception as e:
             logging.exception('[ADMIN] %s', e)
