+++ core/controllers/admin_test.py
@@ -17,7 +17,6 @@
 from __future__ import annotations
 
 import datetime
-import enum
 import logging
 
 from core import feconf
@@ -32,9 +31,6 @@ from core.domain import exp_domain
 from core.domain import exp_services
 from core.domain import fs_services
 from core.domain import opportunity_services
-from core.domain import platform_feature_services
-from core.domain import platform_parameter_domain
-from core.domain import platform_parameter_registry
 from core.domain import question_fetchers
 from core.domain import recommendations_services
 from core.domain import rights_manager
@@ -72,15 +68,6 @@ if MYPY: # pragma: no cover
 
 BOTH_MODERATOR_AND_ADMIN_EMAIL = 'moderator.and.admin@example.com'
 BOTH_MODERATOR_AND_ADMIN_USERNAME = 'moderatorandadm1n'
-
-
-class ParamNames(enum.Enum):
-    """Enum for parameter names."""
-
-    TEST_FEATURE_1 = 'test_feature_1'
-
-
-FeatureStages = platform_parameter_domain.FeatureStages
 
 
 class AdminIntegrationTest(test_utils.GenericTestBase):
@@ -343,81 +330,6 @@ class AdminIntegrationTest(test_utils.GenericTestBase):
                 '/adminhandler', {
                     'action': 'rollback_exploration_to_safe_state',
                     'exp_id': None
-                }, csrf_token=csrf_token)
-
-        self.logout()
-
-    def test_without_feature_name_action_update_feature_flag_is_not_performed(
-        self
-    ) -> None:
-        self.login(self.CURRICULUM_ADMIN_EMAIL, is_super_admin=True)
-        csrf_token = self.get_new_csrf_token()
-
-        assert_raises_regexp_context_manager = self.assertRaisesRegex(
-            Exception,
-            'The \'feature_name\' must be provided when the action is '
-            'update_feature_flag_rules.'
-        )
-        with assert_raises_regexp_context_manager, self.prod_mode_swap:
-            self.post_json(
-                '/adminhandler', {
-                    'action': 'update_feature_flag_rules',
-                    'feature_name': None
-                }, csrf_token=csrf_token)
-
-        self.logout()
-
-    def test_without_new_rules_action_update_feature_flag_is_not_performed(
-        self
-    ) -> None:
-        self.login(self.CURRICULUM_ADMIN_EMAIL, is_super_admin=True)
-        csrf_token = self.get_new_csrf_token()
-
-        assert_raises_regexp_context_manager = self.assertRaisesRegex(
-            Exception,
-            'The \'new_rules\' must be provided when the action is '
-            'update_feature_flag_rules.'
-        )
-        with assert_raises_regexp_context_manager, self.prod_mode_swap:
-            self.post_json(
-                '/adminhandler', {
-                    'action': 'update_feature_flag_rules',
-                    'feature_name': 'new_feature',
-                    'new_rules': None
-                }, csrf_token=csrf_token)
-
-        self.logout()
-
-    def test_without_commit_message_action_update_feature_flag_is_not_performed(
-        self
-    ) -> None:
-        new_rule_dicts = [
-            {
-                'filters': [
-                    {
-                        'type': 'server_mode',
-                        'conditions': [['=', 'dev']]
-                    }
-                ],
-                'value_when_matched': True
-            }
-        ]
-
-        self.login(self.CURRICULUM_ADMIN_EMAIL, is_super_admin=True)
-        csrf_token = self.get_new_csrf_token()
-
-        assert_raises_regexp_context_manager = self.assertRaisesRegex(
-            Exception,
-            'The \'commit_message\' must be provided when the action is '
-            'update_feature_flag_rules.'
-        )
-        with assert_raises_regexp_context_manager, self.prod_mode_swap:
-            self.post_json(
-                '/adminhandler', {
-                    'action': 'update_feature_flag_rules',
-                    'feature_name': 'new_feature',
-                    'new_rules': new_rule_dicts,
-                    'commit_message': None
                 }, csrf_token=csrf_token)
 
         self.logout()
@@ -864,343 +776,6 @@ class AdminIntegrationTest(test_utils.GenericTestBase):
         self.assertEqual(recommendations_services.get_topic_similarity(
             'Art', 'Biology'), 0.2)
 
-        self.logout()
-
-    def test_get_handler_includes_all_feature_flags(self) -> None:
-        self.login(self.CURRICULUM_ADMIN_EMAIL, is_super_admin=True)
-        feature = platform_parameter_registry.Registry.create_feature_flag(
-            ParamNames.TEST_FEATURE_1, 'feature for test.', FeatureStages.DEV)
-
-        feature_list_ctx = self.swap(
-            platform_feature_services, 'ALL_FEATURES_LIST',
-            [ParamNames.TEST_FEATURE_1])
-        feature_set_ctx = self.swap(
-            platform_feature_services, 'ALL_FEATURES_NAMES_SET',
-            set([feature.name]))
-        with feature_list_ctx, feature_set_ctx:
-            response_dict = self.get_json('/adminhandler')
-            self.assertEqual(
-                response_dict['feature_flags'], [feature.to_dict()])
-
-        platform_parameter_registry.Registry.parameter_registry.pop(
-            feature.name)
-        self.logout()
-
-    def test_post_with_flag_changes_updates_feature_flags(self) -> None:
-        self.login(self.CURRICULUM_ADMIN_EMAIL, is_super_admin=True)
-        csrf_token = self.get_new_csrf_token()
-
-        feature = platform_parameter_registry.Registry.create_feature_flag(
-            ParamNames.TEST_FEATURE_1, 'feature for test.', FeatureStages.DEV)
-        new_rule_dicts = [
-            {
-                'filters': [
-                    {
-                        'type': 'server_mode',
-                        'conditions': [['=', 'dev']]
-                    }
-                ],
-                'value_when_matched': True
-            }
-        ]
-
-        feature_list_ctx = self.swap(
-            platform_feature_services, 'ALL_FEATURES_LIST',
-            [ParamNames.TEST_FEATURE_1])
-        feature_set_ctx = self.swap(
-            platform_feature_services, 'ALL_FEATURES_NAMES_SET',
-            set([feature.name]))
-        with feature_list_ctx, feature_set_ctx:
-            self.post_json(
-                '/adminhandler', {
-                    'action': 'update_feature_flag_rules',
-                    'feature_name': feature.name,
-                    'new_rules': new_rule_dicts,
-                    'commit_message': 'test update feature',
-                }, csrf_token=csrf_token)
-
-            rule_dicts = [
-                rule.to_dict() for rule
-                in platform_parameter_registry.Registry.get_platform_parameter(
-                    feature.name).rules
-            ]
-            self.assertEqual(rule_dicts, new_rule_dicts)
-
-        platform_parameter_registry.Registry.parameter_registry.pop(
-            feature.name)
-        self.logout()
-
-    def test_post_flag_changes_correctly_updates_flags_returned_by_getter(
-        self
-    ) -> None:
-        self.login(self.CURRICULUM_ADMIN_EMAIL, is_super_admin=True)
-        csrf_token = self.get_new_csrf_token()
-
-        feature = platform_parameter_registry.Registry.create_feature_flag(
-            ParamNames.TEST_FEATURE_1, 'feature for test.', FeatureStages.DEV)
-        new_rule_dicts = [
-            {
-                'filters': [
-                    {
-                        'type': 'server_mode',
-                        'conditions': [['=', 'dev']]
-                    }
-                ],
-                'value_when_matched': True
-            }
-        ]
-
-        feature_list_ctx = self.swap(
-            platform_feature_services, 'ALL_FEATURES_LIST',
-            [ParamNames.TEST_FEATURE_1])
-        feature_set_ctx = self.swap(
-            platform_feature_services, 'ALL_FEATURES_NAMES_SET',
-            set([feature.name]))
-        with feature_list_ctx, feature_set_ctx:
-            response_dict = self.get_json('/adminhandler')
-            self.assertEqual(
-                response_dict['feature_flags'], [feature.to_dict()])
-
-            self.post_json(
-                '/adminhandler', {
-                    'action': 'update_feature_flag_rules',
-                    'feature_name': feature.name,
-                    'new_rules': new_rule_dicts,
-                    'commit_message': 'test update feature',
-                }, csrf_token=csrf_token)
-
-            response_dict = self.get_json('/adminhandler')
-            rules = response_dict['feature_flags'][0]['rules']
-            self.assertEqual(rules, new_rule_dicts)
-
-        platform_parameter_registry.Registry.parameter_registry.pop(
-            feature.name)
-        self.logout()
-
-    def test_update_flag_rules_with_invalid_rules_returns_400(self) -> None:
-        self.login(self.CURRICULUM_ADMIN_EMAIL, is_super_admin=True)
-        csrf_token = self.get_new_csrf_token()
-
-        feature = platform_parameter_registry.Registry.create_feature_flag(
-            ParamNames.TEST_FEATURE_1, 'feature for test.', FeatureStages.DEV)
-        new_rule_dicts = [
-            {
-                'filters': [
-                    {
-                        'type': 'server_mode',
-                        'conditions': [['=', 'prod']]
-                    }
-                ],
-                'value_when_matched': True
-            }
-        ]
-
-        feature_list_ctx = self.swap(
-            platform_feature_services, 'ALL_FEATURES_LIST',
-            [ParamNames.TEST_FEATURE_1])
-        feature_set_ctx = self.swap(
-            platform_feature_services, 'ALL_FEATURES_NAMES_SET',
-            set([feature.name]))
-        with feature_list_ctx, feature_set_ctx:
-            response = self.post_json(
-                '/adminhandler', {
-                    'action': 'update_feature_flag_rules',
-                    'feature_name': feature.name,
-                    'new_rules': new_rule_dicts,
-                    'commit_message': 'test update feature',
-                },
-                csrf_token=csrf_token,
-                expected_status_int=400
-            )
-            self.assertEqual(
-                response['error'],
-                'Feature in dev stage cannot be enabled in test or production '
-                'environments.')
-
-        platform_parameter_registry.Registry.parameter_registry.pop(
-            feature.name)
-        self.logout()
-
-    def test_update_flag_rules_with_unknown_feature_name_returns_400(
-        self
-    ) -> None:
-        self.login(self.CURRICULUM_ADMIN_EMAIL, is_super_admin=True)
-        csrf_token = self.get_new_csrf_token()
-
-        new_rule_dicts = [
-            {
-                'filters': [
-                    {
-                        'type': 'server_mode',
-                        'conditions': [['=', 'dev']]
-                    }
-                ],
-                'value_when_matched': True
-            }
-        ]
-
-        feature_list_ctx = self.swap(
-            platform_feature_services, 'ALL_FEATURES_LIST', [])
-        feature_set_ctx = self.swap(
-            platform_feature_services, 'ALL_FEATURES_NAMES_SET', set([]))
-        with feature_list_ctx, feature_set_ctx:
-            response = self.post_json(
-                '/adminhandler', {
-                    'action': 'update_feature_flag_rules',
-                    'feature_name': 'test_feature_1',
-                    'new_rules': new_rule_dicts,
-                    'commit_message': 'test update feature',
-                },
-                csrf_token=csrf_token,
-                expected_status_int=400
-            )
-            self.assertEqual(
-                response['error'],
-                'Unknown feature flag: test_feature_1.')
-
-        self.logout()
-
-    def test_update_flag_rules_with_feature_name_of_non_string_type_returns_400(
-        self
-    ) -> None:
-        self.login(self.CURRICULUM_ADMIN_EMAIL, is_super_admin=True)
-        csrf_token = self.get_new_csrf_token()
-
-        response = self.post_json(
-            '/adminhandler', {
-                'action': 'update_feature_flag_rules',
-                'feature_name': 123,
-                'new_rules': [],
-                'commit_message': 'test update feature',
-            },
-            csrf_token=csrf_token,
-            expected_status_int=400
-        )
-        error_msg = (
-            'Schema validation for \'feature_name\' failed: Expected '
-            'string, received 123')
-        self.assertEqual(response['error'], error_msg)
-
-        self.logout()
-
-    def test_update_flag_rules_with_message_of_non_string_type_returns_400(
-        self
-    ) -> None:
-        self.login(self.CURRICULUM_ADMIN_EMAIL, is_super_admin=True)
-        csrf_token = self.get_new_csrf_token()
-
-        response = self.post_json(
-            '/adminhandler', {
-                'action': 'update_feature_flag_rules',
-                'feature_name': 'feature_name',
-                'new_rules': [],
-                'commit_message': 123,
-            },
-            csrf_token=csrf_token,
-            expected_status_int=400
-        )
-        error_msg = (
-            'Schema validation for \'commit_message\' failed: Expected '
-            'string, received 123')
-        self.assertEqual(response['error'], error_msg)
-
-        self.logout()
-
-    def test_update_flag_rules_with_rules_of_non_list_type_returns_400(
-        self
-    ) -> None:
-        self.login(self.CURRICULUM_ADMIN_EMAIL, is_super_admin=True)
-        csrf_token = self.get_new_csrf_token()
-
-        response = self.post_json(
-            '/adminhandler', {
-                'action': 'update_feature_flag_rules',
-                'feature_name': 'feature_name',
-                'new_rules': {},
-                'commit_message': 'test update feature',
-            },
-            csrf_token=csrf_token,
-            expected_status_int=400
-        )
-        error_msg = (
-            'Schema validation for \'new_rules\' failed: Expected list, '
-            'received {}')
-        self.assertEqual(response['error'], error_msg)
-
-        self.logout()
-
-    def test_update_flag_rules_with_rules_of_non_list_of_dict_type_returns_400(
-        self
-    ) -> None:
-        self.login(self.CURRICULUM_ADMIN_EMAIL, is_super_admin=True)
-        csrf_token = self.get_new_csrf_token()
-
-        error_msg = (
-            'Schema validation for \'new_rules\' failed: \'int\' '
-            'object is not subscriptable')
-        response = self.post_json(
-            '/adminhandler', {
-                'action': 'update_feature_flag_rules',
-                'feature_name': 'feature_name',
-                'new_rules': [1, 2],
-                'commit_message': 'test update feature',
-            },
-            csrf_token=csrf_token,
-            expected_status_int=400
-        )
-        self.assertEqual(response['error'], error_msg)
-
-        self.logout()
-
-    def test_update_flag_rules_with_unexpected_exception_returns_500(
-        self
-    ) -> None:
-        self.login(self.CURRICULUM_ADMIN_EMAIL, is_super_admin=True)
-        csrf_token = self.get_new_csrf_token()
-
-        feature = platform_parameter_registry.Registry.create_feature_flag(
-            ParamNames.TEST_FEATURE_1, 'feature for test.', FeatureStages.DEV)
-        new_rule_dicts = [
-            {
-                'filters': [
-                    {
-                        'type': 'server_mode',
-                        'conditions': [['=', 'dev']]
-                    }
-                ],
-                'value_when_matched': True
-            }
-        ]
-
-        feature_list_ctx = self.swap(
-            platform_feature_services, 'ALL_FEATURES_LIST',
-            [ParamNames.TEST_FEATURE_1])
-        feature_set_ctx = self.swap(
-            platform_feature_services, 'ALL_FEATURES_NAMES_SET',
-            set([feature.name]))
-        # Here we use MyPy ignore because we are assigning a None value
-        # where instance of 'PlatformParameter' is expected, and this is
-        # done to Replace the stored instance with None in order to
-        # trigger the unexpected exception during update.
-        platform_parameter_registry.Registry.parameter_registry[
-            feature.name] = None  # type: ignore[assignment]
-        with feature_list_ctx, feature_set_ctx:
-            response = self.post_json(
-                '/adminhandler', {
-                    'action': 'update_feature_flag_rules',
-                    'feature_name': feature.name,
-                    'new_rules': new_rule_dicts,
-                    'commit_message': 'test update feature',
-                },
-                csrf_token=csrf_token,
-                expected_status_int=500
-            )
-            self.assertEqual(
-                response['error'],
-                '\'NoneType\' object has no attribute \'serialize\'')
-
-        platform_parameter_registry.Registry.parameter_registry.pop(
-            feature.name)
         self.logout()
 
     def test_grant_super_admin_privileges(self) -> None:
