+++ core/controllers/release_coordinator.py
@@ -16,12 +16,17 @@
 
 from __future__ import annotations
 
+import logging
+
 from core import feconf
+from core import utils
 from core.controllers import acl_decorators
 from core.controllers import base
 from core.domain import caching_services
+from core.domain import platform_feature_services as feature_services
+from core.domain import platform_parameter_domain as parameter_domain
 
-from typing import Dict
+from typing import Dict, List, Optional, TypedDict
 
 
 class MemoryCacheHandler(
@@ -49,3 +54,121 @@ class MemoryCacheHandler(
     def delete(self) -> None:
         caching_services.flush_memory_caches()
         self.render_json({})
+
+
+class FeatureFlagsNormalizePayloadDict(TypedDict):
+    """Dict representation of FeatureFlag's normalized_payload
+    dictionary.
+    """
+
+    action: Optional[str]
+    feature_name: Optional[str]
+    commit_message: Optional[str]
+    new_rules: Optional[List[parameter_domain.PlatformParameterRule]]
+
+
+class FeatureFlagsHandler(
+    base.BaseHandler[FeatureFlagsNormalizePayloadDict, Dict[str, str]]
+):
+    """Handler for feature-flags."""
+
+    GET_HANDLER_ERROR_RETURN_TYPE = feconf.HANDLER_TYPE_JSON
+    URL_PATH_ARGS_SCHEMAS: Dict[str, str] = {}
+    HANDLER_ARGS_SCHEMAS = {
+        'GET': {},
+        'POST': {
+            'action': {
+                'schema': {
+                    'type': 'basestring',
+                    'choices': [
+                        'update_feature_flag_rules'
+                    ]
+                },
+                # TODO(#13331): Remove default_value when it is confirmed that,
+                # for clearing the search indices of exploration & collection
+                # 'action' field must be provided in the payload.
+                'default_value': None
+            },
+            'feature_name': {
+                'schema': {
+                    'type': 'basestring'
+                },
+                'default_value': None
+            },
+            'commit_message': {
+                'schema': {
+                    'type': 'basestring'
+                },
+                'default_value': None
+            },
+            'new_rules': {
+                'schema': {
+                    'type': 'list',
+                    'items': {
+                        'type': 'object_dict',
+                        'object_class': parameter_domain.PlatformParameterRule
+                    }
+                },
+                'default_value': None
+            },
+        }
+    }
+
+    @acl_decorators.can_access_release_coordinator_page
+    def get(self) -> None:
+        """Handles GET requests."""
+        feature_flag_dicts = feature_services.get_all_feature_flag_dicts()
+        self.render_json({
+            'feature_flags': feature_flag_dicts,
+        })
+
+    @acl_decorators.can_access_release_coordinator_page
+    def post(self) -> None:
+        """Handles POST requests."""
+        assert self.normalized_payload is not None
+        action = self.normalized_payload.get('action')
+        try:
+            # The handler schema defines the possible values of 'action'.
+            # If 'action' has a value other than those defined in the
+            # schema, a Bad Request error will be thrown. Hence, 'action'
+            # must be 'update_feature_flag_rules' if this branch is
+            # executed.
+            assert action == 'update_feature_flag_rules'
+            feature_name = self.normalized_payload.get('feature_name')
+            if feature_name is None:
+                raise Exception(
+                    'The \'feature_name\' must be provided when the action'
+                    ' is update_feature_flag_rules.'
+                )
+            new_rules = self.normalized_payload.get('new_rules')
+            if new_rules is None:
+                raise Exception(
+                    'The \'new_rules\' must be provided when the action'
+                    ' is update_feature_flag_rules.'
+                )
+            commit_message = self.normalized_payload.get('commit_message')
+            if commit_message is None:
+                raise Exception(
+                    'The \'commit_message\' must be provided when the '
+                    'action is update_feature_flag_rules.'
+                )
+
+            assert self.user_id is not None
+            try:
+                feature_services.update_feature_flag_rules(
+                    feature_name, self.user_id, commit_message,
+                    new_rules)
+            except (
+                    utils.ValidationError,
+                    feature_services.FeatureFlagNotFoundException) as e:
+                raise self.InvalidInputException(e)
+
+            new_rule_dicts = [rules.to_dict() for rules in new_rules]
+            logging.info(
+                '[RELEASE-COORDINATOR] %s updated feature %s with new rules: '
+                '%s.' % (self.user_id, feature_name, new_rule_dicts))
+            self.render_json(self.values)
+        except Exception as e:
+            logging.exception('[RELEASE-COORDINATOR] %s', e)
+            self.render_json({'error': str(e)})
+            raise e
