+++ core/templates/domain/admin/admin-backend-api.service.ts
@@ -26,10 +26,6 @@ import {
   CreatorTopicSummary,
   CreatorTopicSummaryBackendDict
 } from 'domain/topic/creator-topic-summary.model';
-import {
-  PlatformParameter,
-  PlatformParameterBackendDict
-} from 'domain/platform_feature/platform-parameter.model';
 import { UrlInterpolationService } from
   'domain/utilities/url-interpolation.service';
 import { Schema } from 'services/schema-default-value.service';
@@ -131,7 +127,6 @@ export interface AdminPageDataBackendDict {
   'viewable_roles': string[];
   'human_readable_roles': HumanReadableRolesBackendResponse;
   'topic_summaries': CreatorTopicSummaryBackendDict[];
-  'feature_flags': PlatformParameterBackendDict[];
 }
 
 export interface AdminPageData {
@@ -144,7 +139,6 @@ export interface AdminPageData {
   viewableRoles: string[];
   humanReadableRoles: HumanReadableRolesBackendResponse;
   topicSummaries: CreatorTopicSummary[];
-  featureFlags: PlatformParameter[];
 }
 
 @Injectable({
@@ -169,11 +163,7 @@ export class AdminBackendApiService {
           humanReadableRoles: response.human_readable_roles,
           viewableRoles: response.viewable_roles,
           topicSummaries: response.topic_summaries.map(
-            CreatorTopicSummary.createFromBackendDict),
-          featureFlags: response.feature_flags.map(
-            dict => PlatformParameter.createFromBackendDict(
-              dict)
-          )
+            CreatorTopicSummary.createFromBackendDict)
         });
       }, errorResponse => {
         reject(errorResponse.error.error);
