+++ core/templates/domain/platform_feature/platform-feature-admin-backend-api.service.spec.ts
@@ -20,8 +20,10 @@ import { HttpClientTestingModule, HttpTestingController } from
   '@angular/common/http/testing';
 import { TestBed, fakeAsync, flushMicrotasks } from '@angular/core/testing';
 
-import { AdminPageConstants } from
-  'pages/admin-page/admin-page.constants';
+import { PlatformParameterFilterType } from
+  'domain/platform_feature/platform-parameter-filter.model';
+import { FeatureStage, PlatformParameter } from
+  'domain/platform_feature/platform-parameter.model';
 import { PlatformFeatureAdminBackendApiService } from
   'domain/platform_feature/platform-feature-admin-backend-api.service';
 import { PlatformParameterRule } from
@@ -30,6 +32,24 @@ import { PlatformParameterRule } from
 describe('PlatformFeatureAdminBackendApiService', () => {
   let featureAdminService: PlatformFeatureAdminBackendApiService;
   let httpTestingController: HttpTestingController;
+  let featureFlagsResponse = {
+    feature_flags: [{
+      name: 'dummy_feature',
+      description: 'this is a dummy feature',
+      data_type: 'bool',
+      rules: [{
+        filters: [{
+          type: PlatformParameterFilterType.ServerMode,
+          conditions: [['=', 'dev'] as [string, string]]
+        }],
+        value_when_matched: true
+      }],
+      rule_schema_version: 1,
+      default_value: false,
+      is_feature: true,
+      feature_stage: FeatureStage.DEV
+    }]
+  };
 
   beforeEach(() => {
     TestBed.configureTestingModule({
@@ -43,6 +63,48 @@ describe('PlatformFeatureAdminBackendApiService', () => {
   afterEach(() => {
     httpTestingController.verify();
   });
+
+  it('should get feature flags data', fakeAsync(() => {
+    let successHandler = jasmine.createSpy('success');
+    let failHandler = jasmine.createSpy('fail');
+
+    let featureFlagsObject = {
+      featureFlags: featureFlagsResponse.feature_flags.map(
+        dict => PlatformParameter.createFromBackendDict(dict))
+    };
+    featureAdminService.getFeatureFlags().then(successHandler, failHandler);
+
+    let req = httpTestingController.expectOne('/feature_flags');
+    expect(req.request.method).toEqual('GET');
+    req.flush(featureFlagsResponse);
+
+    flushMicrotasks();
+    expect(successHandler).toHaveBeenCalledWith(featureFlagsObject);
+    expect(failHandler).not.toHaveBeenCalled();
+  }));
+
+  it('should use rejection handler if backend request fails to fetch features',
+    fakeAsync(() => {
+      let successHandler = jasmine.createSpy('success');
+      let failHandler = jasmine.createSpy('fail');
+
+      featureAdminService.getFeatureFlags().then(
+        successHandler, failHandler);
+
+      var req = httpTestingController.expectOne(
+        '/feature_flags');
+      expect(req.request.method).toEqual('GET');
+
+      req.flush({
+        error: 'Some error in the backend.'
+      }, {
+        status: 500, statusText: 'Internal Server Error'
+      });
+      flushMicrotasks();
+
+      expect(successHandler).not.toHaveBeenCalled();
+      expect(failHandler).toHaveBeenCalledWith('Some error in the backend.');
+    }));
 
   it('should make a request to update the feature flag rules',
     fakeAsync(() => {
@@ -60,8 +122,7 @@ describe('PlatformFeatureAdminBackendApiService', () => {
         'feature_name', 'update message', newRules
       ).then(successHandler, failHandler);
 
-      const req = httpTestingController.expectOne(
-        AdminPageConstants.ADMIN_HANDLER_URL);
+      const req = httpTestingController.expectOne('/feature_flags');
       req.flush({});
       expect(req.request.method).toEqual('POST');
 
@@ -87,8 +148,7 @@ describe('PlatformFeatureAdminBackendApiService', () => {
       'feature_name', 'update message', newRules
     ).then(successHandler, failHandler);
 
-    const req = httpTestingController.expectOne(
-      AdminPageConstants.ADMIN_HANDLER_URL);
+    const req = httpTestingController.expectOne('/feature_flags');
     req.error(new ErrorEvent('Error'));
 
     flushMicrotasks();
