+++ core/templates/domain/platform_feature/platform-feature-admin-backend-api.service.ts
@@ -20,12 +20,22 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { downgradeInjectable } from '@angular/upgrade/static';
 
-import { AdminPageConstants } from
-  'pages/admin-page/admin-page.constants';
 import { PlatformFeatureDomainConstants } from
   'domain/platform_feature/platform-feature-domain.constants';
 import { PlatformParameterRule } from
   'domain/platform_feature/platform-parameter-rule.model';
+import {
+  PlatformParameter,
+  PlatformParameterBackendDict
+} from 'domain/platform_feature/platform-parameter.model';
+
+export interface FeatureFlagsDicts {
+  'feature_flags': PlatformParameterBackendDict[];
+}
+
+export interface FeatureFlagsResponse {
+  featureFlags: PlatformParameter[];
+}
 
 @Injectable({
   providedIn: 'root'
@@ -35,11 +45,27 @@ export class PlatformFeatureAdminBackendApiService {
     private http: HttpClient,
   ) {}
 
+  async getFeatureFlags(): Promise<FeatureFlagsResponse> {
+    return new Promise((resolve, reject) => {
+      this.http.get<FeatureFlagsDicts>(
+        '/feature_flags').toPromise().then(response => {
+        resolve({
+          featureFlags: response.feature_flags.map(
+            dict => PlatformParameter.createFromBackendDict(
+              dict)
+          )
+        });
+      }, errorResponse => {
+        reject(errorResponse.error.error);
+      });
+    });
+  }
+
   async updateFeatureFlag(
       name: string, message: string, newRules: PlatformParameterRule[]):
       Promise<void> {
     await this.http.post(
-      AdminPageConstants.ADMIN_HANDLER_URL,
+      '/feature_flags',
       {
         action: PlatformFeatureDomainConstants.UPDATE_FEATURE_FLAG_RULES_ACTION,
         feature_name: name,
