+++ core/templates/pages/admin-page/config-tab/admin-config-tab.component.spec.ts
@@ -101,8 +101,7 @@ describe('Admin config tab component ', () => {
     humanReadableRoles: {
       MODERATOR: 'moderator'
     },
-    topicSummaries: [],
-    featureFlags: []
+    topicSummaries: []
   };
 
   beforeEach(() => {
