+++ core/templates/pages/admin-page/roles-tab/admin-roles-tab.component.spec.ts
@@ -86,8 +86,7 @@ describe('Admin roles tab component ', function() {
     },
     topicSummaries: [
       sampleTopicSummary
-    ],
-    featureFlags: []
+    ]
   };
 
   beforeEach(() => {
